Add support for querying GTR data to the IODA API
Added an extra parameter to the raw signals API: "sourceParams",
which can be used to pass additional query parameters that might
be specific to a particular data source.

For instance, sourceParams for GTR data allow the requester to
specify which products they want data for by providing the products
as a comma separated list via the sourceParams (e.g.
"WEB_SEARCH,TRANSLATE,GMAIL". Default is to query for WEB_SEARCH
only.

Multiple GTR products are summed to produce a single time series
signal.

Added extra code branches to deal with datasources that do not
support a particular entity type (such as "regions" for GTR). In
these cases, we should now return no data rather than throwing an
error.