entities API: entityType and entityCode are now query parameters
Some query types do not strictly require the entity type or code
so they should not be in-path parameters.

Changed these parameters to be query parameters (e.g. they come
after the '?' in the query URL) and updated Swagger accordingly.

Fixes issue where the API docs insisted that entityType and
entityCode were required parameters, which was confusing for
users and made it impossible to experiment with queries that did
not require these parameters (e.g. `?search=united`).

NOTE: do not merge this commit without also merging the
corresponding one for ioda-ui!