outages API: entityType and entityCode are now query parameters
This allows them to specified as optional in swagger (and
therefore not be labelled as required in our API docs).

Again, don't roll this patch out without the corresponding
update for the UI...

Also, adds a default limit of 2000 alerts or events returned
by the API to try and avoid memory exhaustion inside php.