Add new datasource: gtr-norm
This will automatically normalize the GTR signal for any requested
products, so that they can be easily rendered on a comparable scale.

Most importantly, the BLENDED signal now becomes an average of the
normalized values for the three input products (MAPS, GMAIL and
WEB_SEARCH) so this signal is now also a bit more meaningful than
before.

BLENDED has now been disabled as a valid product to request for
the old "gtr" raw signal, as it doesn't really make sense in
that context now.