Add 'entityFqid' to each timeseries returned by the signals API
This is useful if we want to reconstruct the IODA FQID from an
API result, as we otherwise have no easy way to find out the
other parts of the geo-location hierarchy.

For instance, a query result would previously tell us the
region ID for a queried region, but not the country
or continent code, so we would have to do another entity lookup
to get the rest of the information we need to reconstruct the
full FQID if we need it.

One use case is replaying historical data into watchtower for
testing/debugging purposes. The historical source operates on an
API URL, but most filter plugins expect the data to include a
FQID, so we need to be able to reconstruct it from the JSON
returned by the signal API query.