Add ability to split the time period for a signal into eras
Different eras may have different query parameters (e.g. using
a different geo-location provider, or querying from a different
measurement or datasource). The exact timestamps that bound an
era can be different for each datasource (i.e. we switched over
BGP to a new era on a particular day, but did not make a
corresponding change to AP until 2 weeks later).

The main aim of this work is to transparently handle queries
that span significant changes in the data collection process.