Complete initial implementation of "signal eras"
Era separation logic is tested and appears to be working as
intended.

Suitable queries should be generated for all data sources
depending on the era(s) that the queried time period falls into.

The current era markers defined in the code are mostly
arbitrary, especially as data collection for the new era has only
begun for BGP. These will need to be updated appropriately
before any deployment is undertaken.