Added dev_flag to eras table
The dev_flag allows me to insert eras for testing purposes
that won't accidentally affect the prod API (assuming the
dev_flag is set correctly!).

Added extra sanity check on "from" timestamp when trying to
query eras -- a $from of zero is probably a mistake, so don't
proceed with the eras query.