Combine multiple frames into a single series
We can end up with multiple frames if a tag value is
inconsistent over the queried time period (i.e. an AS name
changed, or there was a problem with the friendly tagging).

Previously, we would just return the "first" frame which
means that any other useable data points in the other frames
were discarded. Now, multiple frames are merged together so
that any null values in the first frame are replaced if there
are non-null values in subsequent frames.

This fixes the issue where certain ASNs (e.g. AS6332) would
not return complete signals if the queried time period spans
May 7 2024/