Revamp entity lookups to improve performance
geoasn lookups with "relatedTo" parameters were taking a very long
time because of all the extra queries to mimic the relationships
and attributes that were not available in the metadata DB.

I've now added them to the DB, so we can leverage those directly
and vastly reduce the number of queries that we have to make. It
also simplifies the code quite a lot, which is a good thing!